move functions into obs class, fix test
