+++ astroquery/mast/core.py
@@ -8,7 +8,6 @@ This the base class for MAST queries.
 from ..query import QueryWithLogin
 from . import utils
 from .auth import MastAuth
-from .cloud import CloudAccess
 from .discovery_portal import PortalAPI
 from .services import ServiceAPI
 
@@ -81,31 +80,6 @@ class MastQueryWithLogin(QueryWithLogin):
         """
         self._auth_obj.logout()
         self._authenticated = False
-
-    def enable_cloud_dataset(self, provider="AWS", profile=None, verbose=True):
-        """
-        Enable downloading public files from S3 instead of MAST.
-        Requires the boto3 library to function.
-
-        Parameters
-        ----------
-        provider : str
-            Which cloud data provider to use.  We may in the future support multiple providers,
-            though at the moment this argument is ignored.
-        profile : str
-            Profile to use to identify yourself to the cloud provider (usually in ~/.aws/config).
-        verbose : bool
-            Default True.
-            Logger to display extra info and warning.
-        """
-
-        self._cloud_connection = CloudAccess(provider, profile, verbose)
-
-    def disable_cloud_dataset(self):
-        """
-        Disables downloading public files from S3 instead of MAST
-        """
-        self._cloud_connection = None
 
     def resolve_object(self, objectname):
         """
